* fixed references for `float64` to `int`

+++ go/day1/sum_test.go
@@ -7,21 +7,21 @@ import (
 )
 
 func TestSum(t *testing.T) {
-	total := lib.Sum([]float64{+1, +1, +1})
+	total := lib.Sum([]int{+1, +1, +1})
 	if total != 3 {
-		t.Errorf("Sum was incorrect, got: %f, want: %d.", total, 3)
+		t.Errorf("Sum was incorrect, got: %d, want: %d.", total, 3)
 	}
 }
 func TestSumTwo(t *testing.T) {
-	total := lib.Sum([]float64{+1, +1, -2})
+	total := lib.Sum([]int{+1, +1, -2})
 	if total != 0 {
-		t.Errorf("Sum was incorrect, got: %f, want: %d.", total, 0)
+		t.Errorf("Sum was incorrect, got: %d, want: %d.", total, 0)
 	}
 }
 
 func TestSumThree(t *testing.T) {
-	total := lib.Sum([]float64{-1, -2, -3})
+	total := lib.Sum([]int{-1, -2, -3})
 	if total != -6 {
-		t.Errorf("Sum was incorrect, got: %f, want: %d.", total, -6)
+		t.Errorf("Sum was incorrect, got: %d, want: %d.", total, -6)
 	}
 }
